requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
PyMuPDF==1.23.8
openai==1.3.7
supabase==2.0.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
schedule==1.2.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
flask==2.3.3
